Rework error reporting
